Apply setAll tests to both adapters
